Move native inference quick start to the main Model Zoo

+++ model_zoo/models/super_resolution/pt_DRUNet/scripts/inference.sh
@@ -1,15 +1,17 @@
 #!/bin/bash
 
 source scripts/config.env
-source scripts/build.sh
+source ../../../scripts/build.sh
 
 if [[ "$1" == "-h" || "$1" == "--help" || $# -lt 2 ]]; then
     echo "Usage: $0 <MODEL_PATH> [<image paths list>]"
     exit 2
 fi
 
-filepaths_list=$CURRENT_MODEL_ZOO_PATH/artifacts/inference/filepaths.list
-rm "$filepaths_list"
+filepaths_list="$(pwd)"/artifacts/inference/filepaths.list
+if [ -f "$filepaths_list" ]; then
+    rm "$filepaths_list"
+fi
 
 for arg in "$@"; do
    if [ "$arg" != "$1" ]
@@ -20,13 +22,12 @@ done
 
 MODEL_PATH="$1"
 
-RESULTS_FOLDER=$CURRENT_MODEL_ZOO_PATH/artifacts/inference/results
+RESULTS_FOLDER="$(pwd)"/artifacts/inference/results
 mkdir -p "$RESULTS_FOLDER"
 
-echo "Run the inference script"
-cd $EVAL_SCRIPT_FOLDER
+cd "$VAI_LIBRARY_SAMPLES_PATH" || exit
 
 EVAL_APP=test_accuracy_$VAI_SAMPLES_POSTFIX
-build_if_not_exists $EVAL_APP
+build_if_not_exists $VAI_LIBRARY_SAMPLES_PATH $EVAL_APP
 
 ./$EVAL_APP $MODEL_PATH $filepaths_list $RESULTS_FOLDER
